+++ src/core/letters-slider/components/letter-slider-rail/letter-slider-rail.tsx
@@ -3,5 +3,5 @@ import { genericShadow } from '../../../shadow/shadow.constants'
 import { LetterSliderRail } from './letter-slider-rail-styled'
 
 export const renderLetterSliderRail = () => (
-  <LetterSliderRail style={genericShadow} />
+  <LetterSliderRail />
 )