+++ src/core/letters-slider/components/letter-slider-thumb/letter-slider-thumb.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
-import { genericShadow } from '../../../shadow/shadow.constants'
-import { LetterSliderThumb } from './letter-slider-thumb.styled'
+import { LetterSliderThumb, LetterSliderThumbInner } from './letter-slider-thumb.styled'
 
-export const renderLetterSliderThumb = () => (
-  <LetterSliderThumb style={genericShadow} />
+export const renderLetterSliderThumb = (name: 'low' | 'high', [ min, max ]: [ number, number ]) => (
+  <LetterSliderThumb>
+    <LetterSliderThumbInner {...{ name, min, max }} />
+  </LetterSliderThumb>
 )