+++ src/core/letters-slider/letter-slider.styled.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components/native'
-import RangeSlider from 'rn-range-slider'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
-import { getThemeProp } from '@core/styled/theme'
-import { TEXT_SIZE } from '../text/text.constants'
+import RangeSlider from 'rn-range-slider'
+import { COLOR } from '@core/colors/colors.constants'
+import { TEXT_SIZE } from '@core/text/text.constants'
 import { SPACING } from '@core/styled'
 
 export const LetterSliderBottomLabelBar = styled.View`
@@ -20,10 +20,10 @@ export const LetterSlider = styled(RangeSlider)`
   width: 100%;
 `
 
-export const LetterSliderLengthIcon = styled(MaterialCommunityIcons).attrs(props => ({
-  name: 'format-letter-spacing',
-  color: getThemeProp('textPrimary')(props),
-  size: TEXT_SIZE.L,
-}))`
-  padding-top: ${SPACING.XXS}px;
+export const LetterSliderWarningIcon = styled(MaterialCommunityIcons).attrs({
+  name: 'alert-circle-outline',
+  color: COLOR.FIRE_BRICK,
+  size: TEXT_SIZE.M,
+})`
+  margin-left: ${SPACING.XXS}px;
 `