+++ src/core/letters-slider/letters-slider.tsx
@@ -2,13 +2,15 @@ import * as React from 'react'
 import { SpacingView } from '@core/styled'
 import { Tx } from '@core/tx'
 import { noop } from '@core/noop/noop'
+import { AlertIcon } from '@core/alert-icon/alert-icon'
+import { useLocalize } from '@core/hooks/use-localize.hook'
 import { renderLetterSliderLabel } from './components/letter-slider-label/letter-slider-label'
 import { renderLetterSliderRail } from './components/letter-slider-rail/letter-slider-rail'
 import { renderLetterSliderThumb } from './components/letter-slider-thumb/letter-slider-thumb'
 import { useLettersSlider } from './hooks/use-letters-slider.hook'
 import { LetterSliderDefaultValues } from './models'
 import {
-  LetterSlider, LetterSliderBottomLabelBar, LetterSliderLengthIcon, LetterSliderTopLabelBar
+  LetterSlider, LetterSliderBottomLabelBar, LetterSliderTopLabelBar
 } from './letter-slider.styled'
 
 interface Props {
@@ -17,18 +19,27 @@ interface Props {
 }
 
 export const LettersSlider = ({ onChange, defaultValues }: Props) => {
-  const { min, max, defaultMin, defaultMax, onValueChanged } = useLettersSlider(onChange, defaultValues)
+  const localize = useLocalize()
+  const { min, max, isWarning, defaultMin, defaultMax, onValueChanged } = useLettersSlider(onChange, defaultValues)
+
+  const _renderLetterSliderThumb = (name: 'low' | 'high') => renderLetterSliderThumb(name, [ min, max ])
 
   return (
-    <SpacingView spacings="XXXS S" type="padding">
+    <SpacingView spacings="XS S" type="padding">
       <LetterSliderTopLabelBar>
-        <LetterSliderLengthIcon />
-        <Tx tx={`${min} - ${max}`} spacings="0 XXS 0 S" bold />
+        <Tx error={isWarning} tx={`${min} - ${max}`} bold />
+
+        <AlertIcon
+          title={localize().alert_letters_slider_title}
+          description={localize().alert_letters_slider_description}
+          isVisible={isWarning}
+          type="error"
+        />
       </LetterSliderTopLabelBar>
 
       <LetterSlider
         step={1}
-        renderThumb={renderLetterSliderThumb}
+        renderThumb={_renderLetterSliderThumb}
         renderRail={renderLetterSliderRail}
         renderRailSelected={noop}
         renderLabel={renderLetterSliderLabel(defaultMin, defaultMax)}