+++ src/dashboard/components/force-index-modal/force-index-modal.styled.ts
@@ -3,6 +3,6 @@ import styled from 'styled-components/native'
 
 export const ForceIndexModalContainer = styled.View`
   flex-flow: row-wrap;
-  justify-content: space-around;
   padding: ${SPACING.L}px 0 ${SPACING.XXL}px;
+  margin-left: ${SPACING.XXS}px;
 `