+++ src/dashboard/components/force-index-modal/force-index-modal.tsx
@@ -10,7 +10,7 @@ interface Props {
 }
 
 export const ForceIndexModal = ({ modalizeRef, handleForceIndex }: Props) => {
-  const INDEXES = [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14 ]
+  const INDEXES = [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15 ]
  
   const onPress = (index: number) => () => {
     handleForceIndex(index)
@@ -21,7 +21,7 @@ export const ForceIndexModal = ({ modalizeRef, handleForceIndex }: Props) => {
     <CustomModalize reference={modalizeRef} adjustToContentHeight>
       <ForceIndexModalContainer>
         {INDEXES.map((index: number, realIndex: number) => (
-          <LetterCard content={String(index)} key={index} onPress={onPress(realIndex)} />
+          <LetterCard content={String(index)} key={index} onPress={onPress(realIndex)} withMargin />
         ))}
       </ForceIndexModalContainer>
     </CustomModalize>