+++ changelog.md
@@ -1,3 +1,16 @@
+## 1.17.0 DASHBOARD REBORN
+### Added
+- Unlocked full potential of search engine by allow to search words up to 15 letters!
+- Alert warning for letters slider
+
+### Changed
+- Dashboar UI and UX
+- More tab UI consistency
+- Charade switch migrated to generic checkbox
+
+### Improved
+- Counter UI
+
 ## 1.16.0 UX CLEAN UP
 ### Improved
 - UI and UX consistency